pyproject.toml
[project]
name = "ml-mapo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "lxml>=6.0.2",
    "pandas>=3.0.1",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pyzmq>=27.1.0",
    "torch>=2.10.0",
    "yfinance>=1.2.0",
]
